refactor google ai stream import
