+++ packages/model-runtime/src/core/streams/index.ts
@@ -1,6 +1,6 @@
 export * from './anthropic';
 export * from './bedrock';
-export * from './google-ai';
+export * from './google';
 export * from './model';
 export * from './ollama';
 export * from './openai';
