+++ packages/model-runtime/src/core/streams/vertex-ai.ts
@@ -3,7 +3,7 @@ import { GroundingSearch } from '@lobechat/types';
 
 import { ModelTokensUsage } from '../../types';
 import { nanoid } from '../../utils/uuid';
-import { type GoogleAIStreamOptions } from './google-ai';
+import { type GoogleAIStreamOptions } from './google';
 import {
   StreamContext,
   StreamProtocolChunk,
