+++ packages/model-runtime/src/providers/google/index.test.ts
@@ -5,7 +5,7 @@ import { ChatStreamPayload } from '@lobechat/types';
 import OpenAI from 'openai';
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
 
-import { LOBE_ERROR_KEY } from '../../core/streams/google-ai';
+import { LOBE_ERROR_KEY } from '../../core/streams';
 import { AgentRuntimeErrorType } from '../../types/error';
 import * as debugStreamModule from '../../utils/debugStream';
 import * as imageToBase64Module from '../../utils/imageToBase64';
